# [improve][broker] added isActive in ManagedCursorImpl